Mesh indices and utility functions are now size_t.

+++ code/3DSConverter.cpp
@@ -399,7 +399,7 @@ void Discreet3DSImporter::ConvertMeshes(aiScene* pcOut)
 				register unsigned int index = aiSplit[p][q];
 				aiFace& face = meshOut->mFaces[q];
 
-				face.mIndices = new unsigned int[3];
+				face.mIndices = new size_t[3];
 				face.mNumIndices = 3;
 
 				for (unsigned int a = 0; a < 3;++a,++base)
